requirements.txt
asgiref==3.2.7
astroid==2.4.1
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2.10
django-allauth==0.42.0
django-cors-headers==3.3.0
django-filter==2.2.0
django-heroku==0.3.1
django-import-export==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupPy==1.14
mccabe==0.6.1
mysqlclient==1.4.6
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.2
psycopg2==2.8.5
pylint==2.5.2
python3-openid==3.1.0
pytz==2018.5
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
whitenoise==5.1.0
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
